requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
flet==0.28.3
flet-desktop==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
repath==0.9.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
